rename app-components to app-common
