+++ src/app-common/locationDetail/location-detail.css
