+++ src/app-common/tests/Footer.test.js
