+++ src/app-common/tests/SearchBox.test.js
